Moved FailureCode class to Sinch.Numbers namespace because it is used across the Numbers API.
