+++ tests/Sinch.Tests/Numbers/HooksTests.cs
@@ -1,5 +1,6 @@
 using System.Text.Json;
 using FluentAssertions;
+using Sinch.Numbers;
 using Sinch.Numbers.Hooks;
 using Xunit;
